refactor: move storage module from htsget-search to htsget-storage
