+++ htsget-http/src/lib.rs
@@ -89,9 +89,9 @@ mod tests {
 
   use htsget_config::storage::local::LocalStorage as ConfigLocalStorage;
   use htsget_config::types::{Headers, JsonResponse, Request, Scheme, Url};
-  use htsget_search::htsget::from_storage::HtsGetFromStorage;
-  use htsget_search::htsget::HtsGet;
-  use htsget_search::storage::local::LocalStorage;
+  use htsget_search::from_storage::HtsGetFromStorage;
+  use htsget_search::HtsGet;
+  use htsget_search::LocalStorage;
   use htsget_test::util::expected_bgzf_eof_data_url;
 
   use super::*;
