+++ htsget-lambda/Cargo.toml
@@ -11,8 +11,8 @@ homepage = "https://github.com/umccr/htsget-rs/blob/main/htsget-lambda/README.md
 repository = "https://github.com/umccr/htsget-rs"
 
 [features]
-s3-storage = ["htsget-config/s3-storage", "htsget-search/s3-storage", "htsget-http/s3-storage", "htsget-test/s3-storage"]
-url-storage = ["htsget-config/url-storage", "htsget-search/url-storage", "htsget-http/url-storage", "htsget-test/url-storage"]
+s3-storage = ["htsget-config/s3-storage", "htsget-search/s3-storage", "htsget-http/s3-storage", "htsget-storage/s3-storage", "htsget-test/s3-storage"]
+url-storage = ["htsget-config/url-storage", "htsget-search/url-storage", "htsget-http/url-storage", "htsget-storage/url-storage", "htsget-test/url-storage"]
 default = []
 
 [dependencies]
@@ -20,10 +20,6 @@ tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 tower-http = { version = "0.5", features = ["cors"] }
 lambda_http = { version = "0.11" }
 lambda_runtime = { version = "0.11" }
-htsget-config = { version = "0.10.0", path = "../htsget-config", default-features = false }
-htsget-search = { version = "0.7.1", path = "../htsget-search", default-features = false }
-htsget-http = { version = "0.4.14", path = "../htsget-http", default-features = false }
-htsget-test = { version = "0.6.1", path = "../htsget-test", features = ["http"], default-features = false }
 serde = { version = "1" }
 serde_json = "1"
 mime = "0.3"
@@ -31,6 +27,12 @@ regex = "1"
 tracing = "0.1"
 tracing-subscriber = "0.3"
 bytes = "1"
+
+htsget-config = { version = "0.10.0", path = "../htsget-config", default-features = false }
+htsget-search = { version = "0.7.1", path = "../htsget-search", default-features = false }
+htsget-http = { version = "0.4.14", path = "../htsget-http", default-features = false }
+htsget-test = { version = "0.6.1", path = "../htsget-test", features = ["http"], default-features = false }
+htsget-storage = { version = "0.1.0", path = "../htsget-storage", default-features = false }
 
 [dev-dependencies]
 async-trait = "0.1"
