+++ htsget-lambda/src/handlers/service_info.rs
@@ -6,7 +6,7 @@ use tracing::instrument;
 
 use htsget_http::get_service_info_json as get_base_service_info_json;
 use htsget_http::Endpoint;
-use htsget_search::htsget::HtsGet;
+use htsget_search::HtsGet;
 
 use crate::handlers::FormatJson;
 use crate::ServiceInfo;
